fix: resolve nozo init bins via package exports
