+++ packages/commitlint-config/package.json
@@ -17,13 +17,14 @@
   "type": "module",
   "exports": {
     "./package.json": "./package.json",
-    ".": "./dist/index.js"
+    ".": "./dist/index.js",
+    "./init": "./dist/init/index.js"
   },
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "bin": {
     "nozo-commitlint": "./dist/cli.js",
-    "nozo-commitlint-init": "./dist/init.js"
+    "nozo-commitlint-init": "./dist/init/bin.js"
   },
   "files": [
     "dist",
