+++ packages/commitlint-config/tsdown.config.ts
@@ -1,11 +1,29 @@
 import { defineConfig } from "tsdown";
 
-export default defineConfig({
-  banner: { js: "#!/usr/bin/env node" },
-  entry: ["src/index.ts", "src/cli.ts", "src/init.ts"],
+const base = defineConfig({
   format: ["esm"],
-  dts: true,
-  clean: true,
   platform: "node",
   outExtensions: () => ({ js: ".js", dts: ".d.ts" }),
 });
+
+export default defineConfig([
+  {
+    ...base,
+    entry: {
+      "index": "src/index.ts",
+      "init/index": "src/init/index.ts",
+    },
+    dts: true,
+    clean: true,
+  },
+  {
+    ...base,
+    entry: {
+      "cli": "src/cli.ts",
+      "init/bin": "src/init/bin.ts",
+    },
+    dts: false,
+    clean: false,
+    banner: { js: "#!/usr/bin/env node" },
+  },
+]);
