+++ packages/eslint-config/src/init/bin.ts
@@ -1,0 +1,4 @@
+import { init } from "./index.js";
+
+await init({ cwd: process.cwd() });
+process.stdout.write("✓ @nozomiishii/eslint-config installed\n");
