+++ packages/lefthook-config/package.json
@@ -16,10 +16,15 @@
   "license": "MIT",
   "author": "Nozomi Ishii",
   "type": "module",
+  "exports": {
+    "./package.json": "./package.json",
+    ".": "./recommended.yaml",
+    "./init": "./dist/init/index.js"
+  },
   "main": "recommended.yaml",
   "bin": {
     "nozo-git-harvest": "./dist/cli.js",
-    "nozo-lefthook-init": "./dist/init.js"
+    "nozo-lefthook-init": "./dist/init/bin.js"
   },
   "files": [
     "dist",
