+++ packages/nozo/package.json
@@ -26,6 +26,7 @@
     "lint": "eslint --max-warnings=0 .",
     "lint:fix": "eslint --max-warnings=0 . --fix",
     "prepublishOnly": "pnpm run build",
+    "test": "vitest run",
     "tsc": "tsc"
   },
   "dependencies": {
@@ -46,7 +47,8 @@
     "@types/which": "3.0.4",
     "eslint": "10.3.0",
     "tsdown": "0.21.10",
-    "typescript": "6.0.3"
+    "typescript": "6.0.3",
+    "vitest": "4.1.5"
   },
   "packageManager": "pnpm@11.0.6",
   "engines": {
