+++ packages/postinstall/package.json
@@ -17,12 +17,13 @@
   "type": "module",
   "exports": {
     "./package.json": "./package.json",
-    ".": "./dist/index.js"
+    ".": "./dist/index.js",
+    "./init": "./dist/init/index.js"
   },
   "types": "./dist/index.d.ts",
   "bin": {
     "postinstall": "./dist/index.js",
-    "nozo-postinstall-init": "./dist/init.js"
+    "nozo-postinstall-init": "./dist/init/bin.js"
   },
   "files": [
     "dist",
