+++ packages/postinstall/src/init/bin.ts
@@ -1,0 +1,4 @@
+import { init } from "./index.js";
+
+await init({ cwd: process.cwd() });
+process.stdout.write("✓ @nozomiishii/postinstall installed\n");
